requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.3
requests==2.26.0
python-telegram-bot==13.7
environs==9.3.3
